requirements.txt
tensorflow
pandas
numpy
dvc
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
typeguard
joblib
types-pyYAML
SCIPY
Flask
Flask-Cors
-e .